RR-0018: Setup Login/Register flow
